move page object to wallet folder
