+++ e2e/pages/wallet/DetectedTokensView.js
